requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.3.0
dnspython==2.4.2
docker-pycreds==0.4.0
email-validator==2.1.0.post1
fastapi==0.104.1
filelock==3.13.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kafka-python==2.0.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
orjson==3.9.10
packaging==23.2
pandas==2.1.3
protobuf==4.25.1
psutil==5.9.6
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
sentry-sdk==1.37.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
wandb==0.16.0
watchfiles==0.21.0
websockets==12.0
aiokafka~=0.8.1